feat: option to force chown of home volume

+++ entrypoint.sh
@@ -8,6 +8,7 @@ USER_HOME=${USER_HOME:-/home/"${USER_NAME}"}
 USER_PASSWD=${USER_PASSWD:-$(openssl passwd -1 -salt $(openssl rand -base64 6) "${USER_NAME}")}
 RDP_SERVER=${RDP_SERVER:-no}
 RUN_AS_ROOT=${RUN_AS_ROOT:-no}
+FORCED_OWNERSHIP=${FORCED_OWNERSHIP:-no}
 
 # Create the user account
 ! grep -q ":${USER_GID}:$" /etc/group && groupadd --gid "${USER_GID}" "${USER_NAME}"
@@ -16,16 +17,25 @@ useradd --shell /bin/bash --uid "${USER_UID}" --gid "${USER_GID}" --password "${
 # Create the user's home if it doesn't exist
 [ ! -d "${USER_HOME}" ] && mkdir -p "${USER_HOME}"
 
-# Take ownership of user's home directory
-if [ "$(stat -c '%u:%g' "${USER_HOME}")" != "${USER_UID}:${USER_GID}" ]; then
-    chown "${USER_UID}":"${USER_GID}" "${USER_HOME}"
+echo "FORCED_OWNERSHIP = $FORCED_OWNERSHIP"
+
+# Take ownership of user's home directory if owned by root or if FORCED_OWNERSHIP is enabled
+OWNER_IDS="$(stat -c '%u:%g' "${USER_HOME}")"
+if [ "${OWNER_IDS}" != "${USER_UID}:${USER_GID}" ]; then
+    if [ "${OWNER_IDS}" == "0:0" ] || [ "${FORCED_OWNERSHIP}" == "yes" ]; then
+        chown -R "${USER_UID}":"${USER_GID}" "${USER_HOME}"
+    else
+        echo "ERROR: User's home '${USER_HOME}' is currently owned by $(stat -c '%U:%G' "${USER_HOME}")"
+        echo "Use option --force-owner to enable user ${USER_NAME} to take ownership"
+        exit 1
+    fi
 fi
 
 # Run in X11 redirection mode (default)
 if echo "${RDP_SERVER}" | grep -q -i -E '^(no|off|false|0)$'; then
 
     # Set up pulseaudio for redirection to UNIX socket
-    cp /root/pulse/client.conf /etc/pulse/client.conf
+    [ -f /root/pulse/client.conf ] && cp /root/pulse/client.conf /etc/pulse/client.conf
 
     # Run in X11 redirection mode as $USER_NAME (default)
     if echo "${RUN_AS_ROOT}" | grep -q -i -E '^(no|off|false|0)$'; then
